pyproject.toml
[tool.poetry]
name = "cybulde-model-training"
version = "0.1.0"
description = "model training module for cyber bullying detection"
authors = ["abhipreets2 <abhipreets2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "~=2.0"
matplotlib = "~=3.7"
hydra-core = "~=1.3"
pydantic = "~=1.10"
fsspec = { version = "~=2023.6", extras = ["gcs"] }
gcsfs = "~=2023.6"
torch = "==2.1.1"
lightning = "==2.1.2"
torchmetrics = "~=1.2"
transformers = { version = "~=4.35", extras = ["torch"] }
mlflow = "==2.5.0"
psycopg2-binary = "~=2.9"
google-cloud-compute = "~=1.13"
google-cloud-secret-manager = "~=2.16"
python-etcd = "~=0.4"
fastapi = "~=0.104"
uvicorn = {version = "~=0.24", extras=["standard"]}

[tool.poetry.group.dev.dependencies]
jupyterlab = "~=4.0"
pytest = "~=7.4"
black = "~=23.7.0"
isort = "~=5.12"
flake8 = "~=6.0"
mypy = "~=1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
